requirements.txt
asgiref==3.11.0
autopep8==2.3.2
django==4.2.17
pycodestyle==2.14.0
pytz==2022.1
sqlparse==0.5.5
Unipath==1.1
dj-database-url==0.5.0
python-dotenv==1.2.1
gunicorn==23.0.0
whitenoise==6.11.0
pillow==12.0.0
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
redis==7.1.0
paho-mqtt==2.1.0
APScheduler==3.11.2ac
opencv-python==4.8.1.78
insightface==0.7.3
numpy==1.26.4
scikit-learn==1.7.2
onnxruntime==1.23.2
